# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for City Service — city request management platform",
        "title": "City Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/contractor/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Contractor sees all requests with status=new",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contractor"
                ],
                "summary": "List available requests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by urgency: low, medium, critical",
                        "name": "urgency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/contractor/requests/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Contractor sees requests assigned to them",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contractor"
                ],
                "summary": "My active requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/contractor/requests/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Contractor views full request details including monitor contact",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contractor"
                ],
                "summary": "Get request details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/contractor/requests/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Contractor closes request with completion data and photo. Status changes: in_progress → done",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contractor"
                ],
                "summary": "Complete request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of days spent",
                        "name": "days_spent",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comment about monitor's work",
                        "name": "comment",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Photo of completed work",
                        "name": "photo",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/contractor/requests/{id}/take": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Contractor takes a request. Status changes: new → in_progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contractor"
                ],
                "summary": "Take request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste your JWT token like: Bearer eyJhbG...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}